api/requirements.txt
amqp==2.5.2
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.2
Django==3.0.5
django-cors-headers==3.2.1
django-templated-mail==1.1.1
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
djoser==2.0.3
importlib-metadata==1.6.0
kombu==4.6.8
Pillow==7.1.1
pkg-resources==0.0.0
pytz==2019.3
redis==3.5.0
sqlparse==0.3.1
vine==1.3.0
zipp==3.1.0
